#59 - Fixed after merge, extracted InMemoryStorageExistsTest from InMemoryStorageTest
